# openapi.yaml
openapi: 3.0.3
info:
  title: Polo Sanca Refrigeration Monitoring API
  description: |
    API for monitoring and managing refrigeration equipment across multiple companies and branches.

    ## Authentication
    - **User Authentication**: JWT tokens via `Authorization: Bearer <token>` header
    - **Equipment Authentication**: API keys via `X-API-Key: <key>` header

    ## Versioning
    - Current version: v1
    - Version included in URL path
    - Previous versions supported for 12 months after new release

  version: 1.0.0
  contact:
    name: Polo Sanca API Support
    email: api@polosanca.com
  license:
    name: Proprietary
servers:
  - url: https://api.polosanca.com/v1
    description: Production server
  - url: https://api-staging.polosanca.com/v1
    description: Staging server
...
paths:
...
  /branches:
    get:
      tags:
        - Branches
      summary: List branches
      description: Get branches for a company
      operationId: listBranches
      security:
        - bearerAuth: []
      parameters:
        - name: company_id
          in: query
          description: Company ID (required for Global Admin, automatic for Company users)
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  branches:
                    type: array
                    items:
                      $ref: '#/components/schemas/Branch'
                  pagination:
                    $ref: '#/components/schemas/Pagination'

    post:
      tags:
        - Branches
      summary: Create branch
      operationId: createBranch
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - company_id
                - name
                - address
              properties:
                company_id:
                  type: string
                name:
                  type: string
                address:
                  type: string
      responses:
        '201':
          description: Branch created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Branch'
  /branches/{branch_id}:
    patch:
      tags:
        - Branches
      summary: Update branch
      operationId: updateBranch
      security:
        - bearerAuth: []
      parameters:
        - name: branch_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                address:
                  type: string
      responses:
        '200':
          description: Branch updated successfully

    delete:
      tags:
        - Branches
      summary: Delete branch
      operationId: deleteBranch
      security:
        - bearerAuth: []
      parameters:
        - name: branch_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Branch deleted successfully
...
  /alerts:
    get:
      tags:
        - Alerts
      summary: List alerts
      description: Get alerts filtered by user permissions
      operationId: listAlerts
      security:
        - bearerAuth: []
      parameters:
        - name: equipment_id
          in: query
          schema:
            type: string
        - name: branch_id
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [active, acknowledged, resolved]
        - name: severity
          in: query
          schema:
            type: string
            enum: [warning, critical]
        - name: start_date
          in: query
          schema:
            type: string
            format: date-time
        - name: end_date
          in: query
          schema:
            type: string
            format: date-time
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  alerts:
                    type: array
                    items:
                      $ref: '#/components/schemas/Alert'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
  /alerts/{alert_id}/acknowledge:
    patch:
      tags:
        - Alerts
      summary: Acknowledge an alert
      operationId: acknowledgeAlert
      security:
        - bearerAuth: []
      parameters:
        - name: alert_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                notes:
                  type: string
      responses:
        '200':
          description: Alert acknowledged successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
...
  /alert-rules:
    get:
      tags:
        - Alert Rules
      summary: List alert rules
      description: Get all alert rules (Global Admin only)
      operationId: listAlertRules
      security:
        - bearerAuth: []
      parameters:
        - name: scope
          in: query
          schema:
            type: string
            enum: [global, company, equipment]
        - name: company_id
          in: query
          schema:
            type: string
        - name: equipment_id
          in: query
          schema:
            type: string
        - name: rule_type
          in: query
          schema:
            type: string
            enum: [temperature_high, temperature_low, pressure_high, pressure_low, door_open, equipment_offline]
        - name: is_active
          in: query
          schema:
            type: boolean
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  alert_rules:
                    type: array
                    items:
                      $ref: '#/components/schemas/AlertRule'
                  pagination:
                    $ref: '#/components/schemas/Pagination'

    post:
      tags:
        - Alert Rules
      summary: Create alert rule
      description: Create a new alert rule (Global Admin only)
      operationId: createAlertRule
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AlertRuleCreate'
      responses:
        '201':
          description: Alert rule created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  message:
                    type: string
  /alert-rules/{rule_id}:
    get:
      tags:
        - Alert Rules
      summary: Get alert rule details
      operationId: getAlertRule
      security:
        - bearerAuth: []
      parameters:
        - name: rule_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertRule'

    patch:
      tags:
        - Alert Rules
      summary: Update alert rule
      operationId: updateAlertRule
      security:
        - bearerAuth: []
      parameters:
        - name: rule_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                threshold_value:
                  type: number
                is_active:
                  type: boolean
                duration_seconds:
                  type: integer
                severity:
                  type: string
                  enum: [warning, critical]
      responses:
        '200':
          description: Alert rule updated successfully

    delete:
      tags:
        - Alert Rules
      summary: Delete alert rule
      operationId: deleteAlertRule
      security:
        - bearerAuth: []
      parameters:
        - name: rule_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Alert rule deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  affected_equipments:
                    type: integer
  /equipments/{equipment_id}/alert-rules:
    get:
      tags:
        - Alert Rules
      summary: Get equipment alert rules
      description: Get all alert rules applicable to specific equipment
      operationId: getEquipmentAlertRules
      security:
        - bearerAuth: []
      parameters:
        - name: equipment_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  equipment_id:
                    type: string
                  applicable_rules:
                    type: array
                    items:
                      $ref: '#/components/schemas/AlertRuleWithPriority'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login

    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Equipment API key for telemetry submission
  schemas:
...
    Branch:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        address:
          type: string
        company_id:
          type: string
        equipment_count:
          type: integer
        created_at:
          type: string
          format: date-time
...
    Alert:
      type: object
      properties:
        id:
          type: string
        equipment_id:
          type: string
        serial:
          type: string
        branch_id:
          type: string
        type:
          type: string
          enum: [temperature_high, temperature_low, pressure_high, pressure_low, door_open, equipment_offline]
        severity:
          type: string
          enum: [warning, critical]
        message:
          type: string
        status:
          type: string
          enum: [active, acknowledged, resolved]
        created_at:
          type: string
          format: date-time
        acknowledged_at:
          type: string
          format: date-time
        acknowledged_by:
          type: string
    AlertRule:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        rule_type:
          type: string
          enum: [temperature_high, temperature_low, pressure_high, pressure_low, door_open, equipment_offline]
        threshold_value:
          type: number
        comparison_operator:
          type: string
          enum: ['>', '<', '=', '>=', '<=']
        duration_seconds:
          type: integer
          description: Condition must persist for this duration before alerting
        severity:
          type: string
          enum: [warning, critical]
        message_template:
          type: string
        scope:
          type: string
          enum: [global, company, equipment]
        scope_id:
          type: string
          description: Company ID or Equipment ID (null for global)
        is_active:
          type: boolean
        created_by:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    AlertRuleCreate:
      type: object
      required:
        - name
        - rule_type
        - threshold_value
        - comparison_operator
        - severity
        - scope
      properties:
        name:
          type: string
        description:
          type: string
        rule_type:
          type: string
          enum: [temperature_high, temperature_low, pressure_high, pressure_low, door_open, equipment_offline]
        threshold_value:
          type: number
        comparison_operator:
          type: string
          enum: ['>', '<', '=', '>=', '<=']
        duration_seconds:
          type: integer
          default: 300
        severity:
          type: string
          enum: [warning, critical]
        message_template:
          type: string
        scope:
          type: string
          enum: [global, company, equipment]
        scope_id:
          type: string
        is_active:
          type: boolean
          default: true
    AlertRuleWithPriority:
      allOf:
        - $ref: '#/components/schemas/AlertRule'
        - type: object
          properties:
            priority:
              type: integer
              description: Rule priority (1=highest)
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
          description: Current page number
        limit:
          type: integer
          description: Items per page
        total:
          type: integer
          description: Total number of items
        total_pages:
          type: integer
          description: Total number of pages
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
...